+++ server/api/configuration/routes/patch_apikey.js
@@ -1,0 +1,19 @@
+import generateApiKey from '../../../lib/generateApiKey';
+
+module.exports = (server) => ({
+  method: 'PATCH',
+  path: '/api/configuration/rotate-apikey',
+  config: {
+    auth: {
+      strategies: [ 'jwt' ],
+      scope: [ 'update:configuration' ]
+    },
+    pre: [
+      server.handlers.managementClient
+    ]
+  },
+  handler: (req, reply) =>
+    generateApiKey(req.storage, req.pre.auth0)
+      .then((hash) => reply({ hash }))
+      .catch(err => reply.error(err))
+});